% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_rows.R
\name{arrange_rows}
\alias{arrange_rows}
\title{Arrange the rows of the file}
\usage{
arrange_rows(entry, challenge = "ilinet")
}
\arguments{
\item{challenge}{one of "ilinet" or "state_ili", indicating which
challenge the submission is for}

\item{file}{A submission file}
}
\value{
An arranged data.frame
}
\description{
Restructure rows based on the template.
}
