% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binned_distribution_quantile.R
\name{binned_distribution_quantile}
\alias{binned_distribution_quantile}
\title{Find quantiles of a binned distribution}
\usage{
binned_distribution_quantile(bins, bin_probabilities, p)
}
\arguments{
\item{bins}{vector of lower endpoints of bins (possibly also including upper
endpoint of last bin), assumed to be in increasing order of bins}

\item{bin_probabilities}{numeric vector of probabilities for all bins,
assumed to be in increasing order of bins}

\item{p}{vector of probabilities giving quantiles to find}
}
\value{
numeric vector of lower endpoints of bins in which the specified
   quantiles fall
}
\description{
Find quantiles of a binned distribution
}
