% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples-to-binned-distributions.R
\name{categorical_samples_to_binned_distribution}
\alias{categorical_samples_to_binned_distribution}
\title{From vector of categorical samples, create data frame of bin probabilities.}
\usage{
categorical_samples_to_binned_distribution(x, bins)
}
\arguments{
\item{x}{character vector of samples}

\item{bins}{character vector of bin values.}
}
\value{
Data frame with columns:
   bin_start_incl: lower endpoints of bins
   bin_end_notincl: upper endpoints of bins
   value: proportion of x falling in bin
}
\description{
From vector of categorical samples, create data frame of bin probabilities.
}
