% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcForecastUtils-package.R
\docType{package}
\name{cdcForecastUtils}
\alias{cdcForecastUtils}
\alias{cdcForecastUtils-package}
\title{cdcForecastUtils - Utility functions for CDC forecasts}
\description{
Utility functions for generating and plotting forecasts from a variety of
   models for infectious disease forecasting challenges run by the CDC.
}
\author{
Graham C. Gibson <gcgibson@umass.edu>, Evan L. Ray <eray@mtholyoke.edu>
}
