% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_start_and_end_to_date_seq.R
\name{date_start_and_end_to_date_seq}
\alias{date_start_and_end_to_date_seq}
\title{Compute the sequence of dates in CDCEpi format from the start date to the end date}
\usage{
date_start_and_end_to_date_seq(date_start, date_end)
}
\arguments{
\item{date_start}{Start of the date sequence}

\item{date_end}{End of the date sequence}
}
\value{
Array of dates between date_start and date_end in CDCEpi format
}
\description{
Compute the sequence of dates in CDCEpi format from the start date to the end date
}
\seealso{
\code{\link{date_start_and_end_to_date_seq}}
}
\keyword{internal}
