% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_and_preprocess_flu_data.R
\name{download_and_preprocess_flu_data}
\alias{download_and_preprocess_flu_data}
\title{Download and preprocess the latest CDC flu data, both national and regional}
\usage{
download_and_preprocess_flu_data(
  latest_year = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{latest_year}{year through which data should be downloaded, defaults to current year}
}
\value{
data frame with latest flu data, preprocessed
}
\description{
Download and preprocess the latest CDC flu data, both national and regional
}
