% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_delphi_data_multi_issue.R
\name{fetch_delphi_data_multi_issue}
\alias{fetch_delphi_data_multi_issue}
\title{Fetch and combine data from the Delphi Epidata API}
\usage{
fetch_delphi_data_multi_issue(
  source = c("fluview", "twitter", "wiki"),
  regions = "nat",
  issues = 1,
  epiweeks_range
)
}
\arguments{
\item{source}{currently, one of "fluview", "twitter", or "wiki"}

\item{regions}{character vector of region codes}

\item{issues}{issues to fetch, e.g. c(201040, 201041, 201042); used only for fluview}

\item{epiweeks_range}{range of epiweeks to fetch, e.g. c(201040, 201530)}
}
\value{
data frame with all results from Epidata
}
\description{
Fetch and combine data from the Delphi Epidata API
}
