% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_csv_from_submission_df.R
\name{generate_csv_from_submission_df}
\alias{generate_csv_from_submission_df}
\title{Create csv file suitable for submission}
\usage{
generate_csv_from_submission_df(submission_df, path)
}
\arguments{
\item{submission_df}{data frame}

\item{path}{full path to csv file}
}
\description{
Create csv file suitable for submission
}
