% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_below_baseline_idx.R
\name{get_below_baseline_idx}
\alias{get_below_baseline_idx}
\title{Compute first week when incidence falls below baseline and remains below for
at least lookahead_length weeks in a row.}
\usage{
get_below_baseline_idx(trajectory, baseline, lookahead_length)
}
\arguments{
\item{trajectory}{a numeric vector with incidence for each time point in a 
season}

\item{baseline}{the threshold that incidence must cross}

\item{lookahead_length}{number of consecutive time points that incidence must
fall below the baseline threshold (default 3)}
}
\value{
the smallest index i such that every entry of
  trajectory[seq(from = i, length = lookahead_length)] is < baseline,
  if such an index exists.  Otherwise, NA_integer_
}
\description{
Compute first week when incidence falls below baseline and remains below for
at least lookahead_length weeks in a row.
}
