% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_date_from_flu_data.R
\name{get_current_date_from_flu_data}
\alias{get_current_date_from_flu_data}
\title{Get current epidemic week from flu data}
\usage{
get_current_date_from_flu_data(flu_data)
}
\arguments{
\item{flu_data}{data frame with year and week columns giving MMWR year and week}
}
\value{
epidemic week in format "2020-EW11"
}
\description{
Get current epidemic week from flu data
}
