% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_time_in_date_seq.R
\name{get_current_time_in_date_seq}
\alias{get_current_time_in_date_seq}
\title{Get index of epidemic week within a sequence of epidemic weeks}
\usage{
get_current_time_in_date_seq(current_time, date_seq)
}
\arguments{
\item{current_time}{character: epidemic week in format "2020-EW10"}

\item{date_seq}{character: sequence of epidemic weeks in format "2020-EW10"}
}
\value{
integer index of current time within date sequence
}
\description{
Get index of epidemic week within a sequence of epidemic weeks
}
