% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline.R
\name{get_ili_baseline}
\alias{get_ili_baseline}
\title{Get the ILI baseline for a combination of region and season}
\usage{
get_ili_baseline(location, year = 2015)
}
\arguments{
\item{location}{a string, either "National", "Region k", or "Regionk" where
k in {1, ..., 10}}

\item{year}{integer year, e.g. 2015}
}
\value{
ili baseline value
}
\description{
Get the ILI baseline for a combination of region and season
}
