% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_required_forecast_horizon.R
\name{get_required_forecast_horizon}
\alias{get_required_forecast_horizon}
\title{How many weeks ahead are required to forecast specified targets}
\usage{
get_required_forecast_horizon(targets, h_max = 0, season_end_ew, cdc_report_ew)
}
\arguments{
\item{targets:}{character vector specifying targets to compute. May include:
"wk ahead", "Below baseline for 3 weeks", "First week below baseline",
"Peak height", "Peak week"}

\item{h_max:}{horizon for week ahead forecasts}

\item{season_end_ew:}{character specification of season end epidemic week,
formatted as "2020-EW20"}

\item{cdc_report_ew:}{character specification of epidemic week corresponding
to most recent cdc report, formatted as "2019-EW40"}
}
\description{
How many weeks ahead are required to forecast specified targets
}
