% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_location_code.R
\name{get_standard_location_code}
\alias{get_standard_location_code}
\title{Convert a location name to standardized name as given here:
https://github.com/cmu-delphi/delphi-epidata/blob/master/src/acquisition/fluview/fluview_locations.py}
\usage{
get_standard_location_code(location)
}
\arguments{
\item{location:}{a character vector of non-standardized location names}
}
\value{
a character vector of standardized location names
}
\description{
Convert a location name to standardized name as given here:
https://github.com/cmu-delphi/delphi-epidata/blob/master/src/acquisition/fluview/fluview_locations.py
}
