% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets_regions.R
\name{get_targets_regions}
\alias{get_targets_regions}
\title{Save submitted targets/locations of a file}
\usage{
get_targets_regions(sanitized_file)
}
\arguments{
\item{file}{A csv entry file path}
}
\value{
A data.frame of a set of submitted targets/locations of a model
}
\description{
This function will check to make sure the structure is correct and that
the forecast probabilities are non-negative and sum to a value between
0.9 and 1.1. For binary targets, this function will check if the forecast
provavilities are between 0 and 1.
}
