% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_from_start_of_season.R
\name{get_time_from_start_of_season}
\alias{get_time_from_start_of_season}
\title{Compute the time from the start of season}
\usage{
get_time_from_start_of_season(season_start, current_time)
}
\arguments{
\item{current_time}{End of the date sequence}

\item{date_start}{Start of the date sequence}
}
\value{
integer of time between date_start and current_time
}
\description{
Compute the time from the start of season
}
\seealso{
\code{\link{get_time_from_start_of_season}}
}
\keyword{internal}
