% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viz_from_submission_df.R
\name{get_viz_from_submission_df}
\alias{get_viz_from_submission_df}
\title{Compute the sequence of dates in CDCEpi format from the start date to the end date}
\usage{
get_viz_from_submission_df(submission_df)
}
\arguments{
\item{submission_df}{Full submission file as dataframe from either inf code or read_entry}
}
\value{
A plot list unique to targets
}
\description{
Compute the sequence of dates in CDCEpi format from the start date to the end date
}
\seealso{
\code{\link{get_viz_from_submission_df}}
}
\keyword{internal}
