% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_logscore.R
\name{internal_logscore}
\alias{internal_logscore}
\title{Create a data frame containing a row with a log score column for use during a season}
\usage{
internal_logscore(file, challenge, targets, start_ew)
}
\arguments{
\item{file}{A file path}

\item{challenge}{one of "ilinet" or "state_ili", indicating which
challenge the submission is for}

\item{targets}{a set of week ahead targets to score}

\item{start_ew}{a starting week of a challenge}
}
\value{
data.frame with log scores
}
\description{
Create a data frame containing a row with a log score column for use during a season
}
