% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories_and_intervals.R
\name{plot_trajectories_and_intervals}
\alias{plot_trajectories_and_intervals}
\title{Plot trajectories and intervals}
\usage{
plot_trajectories_and_intervals(
  flu_data,
  target_variable,
  trajectories_by_location,
  submission,
  season_start_ew,
  season_end_ew,
  cdc_report_ew
)
}
\arguments{
\item{flu_data}{A dataframe housing the flu data from the result of 
download_and_preprocess...}

\item{target_variable}{}

\item{trajectories_by_location}{A series of trajectories (nsim*h) per location}

\item{submission}{the submission dataframe}
}
\value{
A plot
}
\description{
Plot trajectories and intervals
}
