% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_entry.R
\name{sanitize_entry}
\alias{sanitize_entry}
\title{Sanitize a csv entry file}
\usage{
sanitize_entry(entry)
}
\arguments{
\item{entry}{A data.frame of a csv entry}
}
\value{
A correctly formatted data.frame
}
\description{
This function re-format entry files for minor formatting issues.
}
