% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_targets.R
\name{standardize_targets}
\alias{standardize_targets}
\title{Convert targets to standardize form}
\usage{
standardize_targets(targets)
}
\arguments{
\item{targets:}{character vector of targets}
}
\value{
character vector of valid targets
}
\description{
Convert targets to standardize form
}
