% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories_to_binned_distributions.R
\name{trajectories_to_binned_distributions}
\alias{trajectories_to_binned_distributions}
\title{Convert a matrix of sampled trajectories to binned distributions for
short-term and/or seasonal targets.}
\usage{
trajectories_to_binned_distributions(
  trajectories,
  h_max,
  bins,
  season_start_ew,
  season_end_ew,
  cdc_report_ew
)
}
\arguments{
\item{trajectories}{matrix of simulated trajectories.  Each row is one
simulated trajectory, each column is one time point.  Must contain all
time points needed to calculate seasonal and short-term targets}

\item{h_max}{largest horizon for short-term targets}

\item{bins:}{named list of bin start and end points.  For example:
list(
  "season peak incidence" = c(seq(from = 0.0, to = 25.0, by = 0.1), 100.0),
  "season peak timing" = c(seq(from = 40, to = 53), seq(from = 1, to = 20)),
  "onset timing" = c(as.character(c(seq(from = 40, to = 53), seq(from = 1, to = 20))), "none")}

\item{season_start_ew:}{character specification of season start epidemic
week, formatted as "2019-EW40"}

\item{season_end_ew:}{character specification of season end epidemic week,
formatted as "2020-EW20"}

\item{cdc_report_ew:}{character specification of epidemic week corresponding
to most recent cdc report, formatted as "2019-EW40"}
}
\value{
data frame with columns:
   target: with values coming from names(bins)
   type: populated with "Bin"
   bin_start_incl: lower endpoints of bins
   bin_end_notincl: upper endpoints of bins
   value: proportion of trajectories falling in bin
}
\description{
Convert a matrix of sampled trajectories to binned distributions for
short-term and/or seasonal targets.
}
