% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_filename.R
\name{verify_filename}
\alias{verify_filename}
\title{Verifies appropriate filename structure}
\usage{
verify_filename(filename, challenge)
}
\arguments{
\item{filename}{a string representing a forecast entry filename}

\item{challenge}{string indicating which project the submission is for}
}
\value{

}
\description{
Verifies appropriate filename structure
}
