% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_targets.R
\name{verify_targets}
\alias{verify_targets}
\title{Verify targets are correct}
\usage{
verify_targets(entry, challenge = "ilinet")
}
\arguments{
\item{entry}{An entry data.frame}

\item{challenge}{one of "ilinet" or "state_ili", indicating which
challenge the submission is for}
}
\value{
Invisibly returns TRUE if successful
}
\description{
The targets depend on the challenge, so this will verify targets are correct
for all targets based on the challenge the submission is for
}
\examples{
verify_targets(full_entry_new)
}
\seealso{
\code{\link{verify_entry}}
}
\keyword{internal}
