% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sanitized_file.R
\name{write_sanitized_file_c}
\alias{write_sanitized_file_c}
\title{Write a sanitized file}
\usage{
write_sanitized_file_c(file, challenge = "ilinet")
}
\arguments{
\item{file}{A csv file path}

\item{challenge}{Either "ilinet" or "state_ili", indicating challenge the submission is for.}
}
\value{
A reformatted csv file
}
\description{
This function reads in the csv file and arranges it for consistency and overwrite.
}
