% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_start_and_end_to_date_seq.R
\name{date_start_and_end_to_date_seq}
\alias{date_start_and_end_to_date_seq}
\title{Get sequence of epidemic weeks between start and end points}
\usage{
date_start_and_end_to_date_seq(date_start, date_end)
}
\arguments{
\item{date_start}{character: epidemic week in format "2020-EW10"}

\item{date_end}{character: epidemic week in format "2020-EW40"}
}
\value{
character: sequence of epidemic weeks from date_start to date_end, inclusive
}
\description{
Get sequence of epidemic weeks between start and end points
}
