% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_targets_regions.R
\name{get_targets_regions}
\alias{get_targets_regions}
\title{Save submitted targets/locations of a file}
\usage{
get_targets_regions(sanitized_file)
}
\arguments{
\item{sanitized_file}{A csv entry file path}
}
\value{
A data.frame of a set of submitted targets/locations of a model
}
\description{
This function will retrieve information from an entry file about what 
locations and targets were submitted for a particular forecast due date.
}
