% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories_and_intervals.R
\name{plot_trajectories_and_intervals}
\alias{plot_trajectories_and_intervals}
\title{Plot trajectories and intervals}
\usage{
plot_trajectories_and_intervals(
  flu_data,
  target_variable,
  trajectories_by_location,
  submission,
  season_start_ew,
  season_end_ew,
  cdc_report_ew,
  ylim,
  lookback_length = 30,
  plot = TRUE
)
}
\arguments{
\item{flu_data}{A dataframe housing the flu data from the result of 
download_and_preprocess...}

\item{target_variable}{variable to be plotted}

\item{trajectories_by_location}{A series of trajectories (nsim*h) per location}

\item{submission}{the submission dataframe}

\item{season_start_ew}{epidemic week starting the season}

\item{season_end_ew}{epidemic week ending the season}

\item{cdc_report_ew}{epidemic week of the CDC data}

\item{ylim}{numeric vector specifying lower and upper limits of y axis; optional}

\item{lookback_length}{integer number of weeks of reported ILINet data to
include}

\item{plot}{boolean;}
}
\value{
logical. If TRUE (default), the plot is printed.
}
\description{
Plot trajectories and intervals
}
