% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sanitized_file.R
\name{write_sanitized_file}
\alias{write_sanitized_file}
\title{Write a sanitized file}
\usage{
write_sanitized_file(file, challenge = "ilinet")
}
\arguments{
\item{file}{A csv file path}
}
\value{
An arranged data.frame
}
\description{
This function reads in the csv file and arranges it for consistency.
}
